requirements.txt
Django==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
Pillow==9.5.0
django-extensions==3.2.3
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
